// src/api/models/artifact-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ArtifactTO
 */
export interface ArtifactTO {
    /**
     * 
     * @type {string}
     * @memberof ArtifactTO
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactTO
     */
    repositoryId: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactTO
     */
    name: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactTO
     */
    description: any;
    /**
     * 
     * @type {Date}
     * @memberof ArtifactTO
     */
    createdDate?: any;
    /**
     * 
     * @type {Date}
     * @memberof ArtifactTO
     */
    updatedDate?: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactTO
     */
    svgPreview?: any;
    /**
     * 
     * @type {string}
     * @memberof ArtifactTO
     */
    fileType: any;
}